fix: 修复引用redux-persist导致刷新后持续loading的bug
